Initial commit of the separate Blackboard API Client
Previously part of github.com/sanjacob/BlackboardSync

This is a slightly improved version of the version currently on the main
branch at the upstream. First, the API module has been adapted to use
the tiny-api-client PyPI package. Second, two new modules have been
introduced: api_extended and filters, which work together to allow
developers to quickly perform compound operations with the API and even
directly filter the results with a blacklist and a whitelist. For
instance, fetching user courses used to be a two-step operation, and the
extended api class has made this into a single call, which takes in a
course filter as a parameter.

This library existing independently might improve the development cycle
of BlackboardSync, as well as allow other developers to use the library
on its own, and add to it.

Special thanks to Cainan Mahar who wrote the BBAvailable enum in the
BlackboardSync repo.

Co-authored-by: Cainan Mahar <52504207+cainanmahar@users.noreply.github.com>